+++ src/index.tsx
@@ -99,7 +99,17 @@ export {
     NavigationBarActionGroup,
     NavigationBarAction,
 } from './navigation-bar';
-export {Logo, MovistarLogo, VivoLogo, O2Logo, O2NewLogo, TelefonicaLogo, BlauLogo, TuLogo} from './logo';
+export {
+    Logo,
+    MovistarLogo,
+    VivoLogo,
+    O2Logo,
+    O2NewLogo,
+    TelefonicaLogo,
+    BlauLogo,
+    TuLogo,
+    EsimflagLogo,
+} from './logo';
 export {default as Image} from './image';
 export {default as Chip} from './chip';
 export {default as Video} from './video';
@@ -204,6 +214,7 @@ export {
     TELEFONICA_SKIN,
     BLAU_SKIN,
     TU_SKIN,
+    ESIMFLAG_SKIN,
 } from './skins/constants';
 export {getSkinByName} from './skins/utils';
 export {getVivoSkin} from './skins/vivo';
@@ -214,6 +225,7 @@ export {getO2NewSkin} from './skins/o2-new';
 export {getTelefonicaSkin} from './skins/telefonica';
 export {getBlauSkin} from './skins/blau';
 export {getTuSkin} from './skins/tu';
+export {getEsimflagSkin} from './skins/esimflag';
 export {palette as vivoPalette} from './skins/vivo';
 export {palette as movistarPalette} from './skins/movistar';
 export {palette as o2Palette} from './skins/o2';
@@ -221,6 +233,7 @@ export {palette as o2NewPalette} from './skins/o2-new';
 export {palette as telefonicaPalette} from './skins/telefonica';
 export {palette as blauPalette} from './skins/blau';
 export {palette as tuPalette} from './skins/tu';
+export {palette as esimFlagPalette} from './skins/esimflag';
 
 export type {TrackingEvent, IconProps, DataAttributes} from './utils/types';
 export type {RegionCode} from './utils/region-code';